+++ CHANGELOG.md
@@ -3,5 +3,9 @@ All notable changes to the "scs" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Implement syntax highlight for scs
+## [0.2.0] - Current
+- Implement hover events
+- Implement diagnostic issues
+
+## [0.1.0]
+- Implement syntax highlight for SCs